Fix compiler warnings on 32-bit

+++ src/assign.c
@@ -2819,7 +2819,7 @@ static void assign_ptr(void) {
   if (!ateol[*basicvars.current]) error(ERR_SYNTAX);
   fileio_setptr(handle, newplace);
   } else if (*basicvars.current=='(') {
-    int64 newptr;
+    size_t newptr;
     stackitem topitem;
     basicarray *descriptor;
     variable *vp;
